requirements.txt
bcrypt==4.1.3
black==24.4.2
blinker==1.8.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
flask-swagger-ui==4.11.1
importlib_resources==6.4.0
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.12.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
ordered-set==4.1.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
rich==13.7.1
typing_extensions==4.12.2
Werkzeug==3.0.3
wrapt==1.16.0
